+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## [Unreleased]
 
+## [0.2.0] - 2026-04-02
+
+- Add `TNumber` as alias of `TrustedNumber` class
+- Add factory `TrustedNumber.factory(base)`
+- Add more tests
+
 ## [0.1.1] - 2026-03-31
 
 - Add MPL 2.0 license